+++ nextjs/features/auth/styles/login-form.module.scss
@@ -1,0 +1,150 @@
+.loginContainer {
+  display: grid;
+  .loginContent {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    overflow: scroll;
+    .loginFormContent {
+
+      .loginFormContentSubTitle {
+        font-size: 14px;
+        font-weight: bold;
+        line-height: 14px;
+        letter-spacing: 0.21666669845581055px;
+        text-align: center;
+        font-family: 'October_Devanagari', sans-serif;
+      }
+
+      .loginTextFormContainer {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        .textForm {
+          font-size: 16px;
+          font-weight: bold;
+          margin-bottom: 1.5rem;
+          max-width: 20rem;
+          width: 100%;
+          &:first-child {
+            margin-top: 3rem;
+          }
+          .warningLabel {
+            font-size: 11px;
+          }
+        }
+
+        .loginFormPasswordContainer {
+          position: relative;
+
+          .eyeIconSlash {
+            position: absolute;
+            right: 10px;
+            top: 50%;
+            width: 25px;
+            height: 25px;
+            transform: translateY(-50%);
+            cursor: pointer;
+            color: #D9D9D9;
+          }
+  
+          .eyeIcon {
+            position: absolute;
+            right: 10px;
+            top: 50%;
+            width: 25px;
+            height: 25px;
+            transform: translateY(-50%);
+            cursor: pointer;
+            color: #D9D9D9;
+          }
+        }
+
+        .createButtonContainer {
+          width: 100%;
+          max-width: 20rem;
+          margin-bottom: 2rem;
+          .createButton {
+            background-color: #69beef;
+            padding: 1rem 2.5rem;
+            color: #fff;
+            border-radius: 0.5rem;
+            width: 100%;
+          }
+        }
+
+        .separateLoginContainer {
+          .separateLoginText {
+            font-size: 11px;
+          }
+          &:not(:first-child) {
+            margin-top: 0.5rem;
+          }
+        }
+
+        .googleLoginContainer {
+          margin: 2rem;
+          width: 100%;
+          max-width: 15rem;
+          border-radius: 0.5rem;
+          .googleLoginButton {
+            width: 100%;
+            display: flex;
+            padding: 1rem 2.5rem 1rem 1.5rem;
+            background-color: #fff;
+          }
+          .googleLoginText {
+            margin-left: 2rem;
+          }
+        }
+      }
+    }
+  }
+}
+
+// 1025px以上（PC）
+@media (min-width: 1024px) {
+  .loginContainer {
+    grid-auto-flow: column;
+    grid-template-columns: 1fr 1fr;
+    height: calc(100vh);
+  }
+  .loginFormContent {
+    width: 80%;
+  }
+}
+
+// 1024px未満（タブレット）
+@media (min-width: 768px) AND (max-width: 1023px) {
+  .loginContainer {
+    height: calc(100vh);
+  }
+  .loginImageContainer {
+    display: none;
+  }
+  .loginFormContent {
+    width: 80%;
+  }
+}
+
+// 1024px未満（タブレット）
+@media (max-width: 767px) {
+  .loginContainer {
+    height: calc(100vh);
+  }
+  .loginImageContainer {
+    display: none;
+  }
+  .loginFormContent {
+    width: 100%;
+  }
+}
+
+.errorMailMessage {
+  color: red;
+}
+
+.errorPasswordMessage {
+  color: red;
+}
