--- nextjs/features/auth/styles/sign-in-form.module.scss
@@ -1,106 +1,0 @@
-.spinner {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-.errorMailMessage {
-  color: red;
-  font-family: "Noto Sans JP", sans-serif;
-}
-
-.errorPasswordMessage {
-  color: red;
-  font-family: "Noto Sans JP", sans-serif;
-}
-
-.signInFormContainerModal {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.signInFormContainer {
-  display: flex;
-
-  .signInFormContainerImageWrapper {
-    flex: 1;
-    position: relative;
-    height: 100%;
-
-    .signInFormContainerImage {
-      width: 100%;
-      height: 100%;
-      object-fit: fill;
-    }
-  }
-
-  .signInFormCntentContainer {
-    flex: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    .signInFormContentWrapper {
-      max-width: 448px;
-      padding: 2rem;
-
-      .signInFormContentTitle {
-        font-size: 40px;
-        font-weight: bold;
-        line-height: 40px;
-        letter-spacing: 0.21666669845581055px;
-        text-align: center;
-        font-family: 'October_Devanagari', sans-serif;
-        margin-bottom: 0.5rem;
-      }
-
-      .signInFormContentSubTitle {
-        font-size: 14px;
-        font-weight: bold;
-        line-height: 14px;
-        letter-spacing: 0.21666669845581055px;
-        text-align: center;
-        font-family: 'October_Devanagari', sans-serif;
-        margin-bottom: 2rem;
-      }
-    }
-
-    .signInFormContentInputContainer {
-      display: flex;
-      flex-direction: column;
-      gap: 2rem;
-      width: 260px;
-      margin-left: auto;
-      margin-right: auto;
-
-      .signInFormContentPasswordContainer {
-        position: relative;
-
-        .eyeIconSlash {
-          position: absolute;
-          right: 10px;
-          top: 50%;
-          width: 25px;
-          height: 25px;
-          transform: translateY(-50%);
-          cursor: pointer;
-          color: #D9D9D9;
-        }
-
-        .eyeIcon {
-          position: absolute;
-          right: 10px;
-          top: 50%;
-          width: 25px;
-          height: 25px;
-          transform: translateY(-50%);
-          cursor: pointer;
-          color: #D9D9D9;
-        }
-      }
-    }
-  }
-}
